+++ openedx_authz/__init__.py
@@ -4,6 +4,6 @@ Open edX AuthZ provides the architecture and foundations of the authorization fr
 
 import os
 
-__version__ = "0.9.1"
+__version__ = "0.10.0"
 
 ROOT_DIRECTORY = os.path.dirname(os.path.abspath(__file__))